@@ -1,13 +1,67 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Item_Trading_App_REST_API.Models.Base;
+using MapsterMapper;
+using Microsoft.AspNetCore.Mvc;
 using System.Linq;
 
-namespace Item_Trading_App_REST_API.Controllers
+namespace Item_Trading_App_REST_API.Controllers;
+
+public class BaseController : Controller
 {
-    public class BaseController : Controller
+    protected readonly IMapper _mapper;
+
+    public BaseController(IMapper mapper)
     {
-        protected string UserId
+        _mapper = mapper;
+    }
+
+    protected string UserId
+    {
+        get => User.Claims.First(c => Equals(c.Type, "id"))?.Value;
+    }
+
+    protected R AdaptToType<T, R>(T request, params (string, object)[] parameters)
+    {
+        var builder = _mapper.From(request);
+
+        if (parameters is not null)
+            foreach (var parameter in parameters)
+            {
+                builder = builder.AddParameters(parameter.Item1, parameter.Item2);
+            }
+
+        return builder.AdaptToType<R>();
+    }
+
+    protected ObjectResult MapResult<InputType, SucceededType, FailedType>(InputType result)
+        where InputType : BaseResult
+        where SucceededType : class
+        where FailedType : class
+    {
+        return result.Success ?
+            Ok(_mapper.From(result).AdaptToType<SucceededType>()) :
+            BadRequest(_mapper.From(result).AdaptToType<FailedType>());
+    }
+
+    protected ObjectResult MapResult<InputType, SucceededType, FailedType>(InputType result, params (string, object)[] parameters)
+        where InputType : BaseResult
+        where SucceededType : class
+        where FailedType : class
+    {
+        if (result.Success)
         {
-            get => User.Claims.First(c => Equals(c.Type, "id"))?.Value;
+            var builder = _mapper.From(result);
+
+            if (parameters is not null)
+                foreach (var parameter in parameters)
+                {
+                    builder = builder.AddParameters(parameter.Item1, parameter.Item2);
+                }
+
+            return Ok(builder.AdaptToType<SucceededType>());
+        }
+        else
+        {
+            return BadRequest(_mapper.From(result).AdaptToType<FailedType>());
         }
     }
 }
