@@ -2,87 +2,44 @@
 using Item_Trading_App_Contracts.Requests.Wallet;
 using Item_Trading_App_Contracts.Responses.Base;
 using Item_Trading_App_Contracts.Responses.Wallet;
+using Item_Trading_App_REST_API.Models.Wallet;
 using Item_Trading_App_REST_API.Services.Wallet;
+using MapsterMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 
-namespace Item_Trading_App_REST_API.Controllers
+namespace Item_Trading_App_REST_API.Controllers;
+
+[Authorize]
+public class WalletController : BaseController
 {
-    [Authorize]
-    public class WalletController : BaseController
+    private readonly IWalletService _walletService;
+
+    public WalletController(IWalletService walletService, IMapper mapper) : base(mapper)
     {
-        private readonly IWalletService _walletService;
+        _walletService = walletService;
+    }
 
-        public WalletController(IWalletService walletService)
-        {
-            _walletService = walletService;
-        }
+    [HttpGet(Endpoints.Wallet.Get)]
+    public async Task<IActionResult> Get()
+    {
+        var result = await _walletService.GetWalletAsync(UserId);
 
-        [HttpGet(Endpoints.Wallet.Get)]
-        public async Task<IActionResult> Get()
-        {
-            string userId = UserId;
+        return MapResult<WalletResult, WalletSuccessResponse, FailedResponse>(result);
+    }
 
-            if(string.IsNullOrEmpty(userId))
+    [HttpPatch(Endpoints.Wallet.Update)]
+    public async Task<IActionResult> Update([FromBody] UpdateWalletRequest request)
+    {
+        if (request is null)
+            return BadRequest(new FailedResponse
             {
-                return BadRequest(new FailedResponse
-                {
-                    Errors = new[] { "User id not found" }
-                });
-            }
-
-            var wallet = await _walletService.GetWalletAsync(userId);
-
-            if(wallet == null)
-            {
-                return BadRequest(new FailedResponse 
-                {
-                    Errors = new[] { "Something went wrong" }
-                });
-            }
-
-            if(!wallet.Success)
-            {
-                return BadRequest(new FailedResponse
-                {
-                    Errors = wallet.Errors
-                });
-            }
-
-            return Ok(new WalletSuccessResponse
-            {
-                Cash = wallet.Cash
+                Errors = new[] { "Something went wrong" }
             });
-        }
 
-        [HttpPatch(Endpoints.Wallet.Update)]
-        public async Task<IActionResult> Update([FromBody] UpdateWalletRequest request)
-        {
-            var userId = UserId;
+        var result = await _walletService.UpdateWalletAsync(AdaptToType<UpdateWalletRequest, UpdateWallet>(request, ("userId", UserId)));
 
-            if (string.IsNullOrEmpty(userId) || request == null)
-            {
-                return BadRequest(new FailedResponse
-                {
-                    Errors = new[] { "Something went wrong" }
-                });
-            }
-
-            var wallet = await _walletService.UpdateWalletAsync(userId, request.Quantity);
-
-            if(!wallet.Success)
-            {
-                return BadRequest(new FailedResponse
-                {
-                    Errors = wallet.Errors
-                });
-            }
-
-            return Ok(new UpdateWalletSuccessResponse
-            {
-                Amount = wallet.Cash
-            });
-        }
+        return MapResult<WalletResult, UpdateWalletSuccessResponse, FailedResponse>(result);
     }
 }
