@@ -1,15 +1,14 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Item_Trading_App_REST_API.Entities
+namespace Item_Trading_App_REST_API.Entities;
+
+public class Item
 {
-    public class Item
-    {
-        [Key]
-        public string ItemId { get; set; }
+    [Key]
+    public string ItemId { get; set; }
 
-        [Required]
-        public string Name { get; set; }
+    [Required]
+    public string Name { get; set; }
 
-        public string Description { get; set; }
-    }
+    public string Description { get; set; }
 }
