@@ -2,27 +2,26 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Item_Trading_App_REST_API.Entities
+namespace Item_Trading_App_REST_API.Entities;
+
+public class RefreshToken
 {
-    public class RefreshToken
-    {
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public string Token { get; set; }
+    [Key]
+    [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+    public string Token { get; set; }
 
-        public string JwtId { get; set; }
+    public string JwtId { get; set; }
 
-        public DateTime CreationDate { get; set; }
+    public DateTime CreationDate { get; set; }
 
-        public DateTime ExpiryDate { get; set; }
+    public DateTime ExpiryDate { get; set; }
 
-        public bool Used { get; set; }
+    public bool Used { get; set; }
 
-        public bool Invalidated { get; set; }
+    public bool Invalidated { get; set; }
 
-        public string UserId { get; set; }
+    public string UserId { get; set; }
 
-        [ForeignKey(nameof(UserId))]
-        public User User { get; set; }
-    }
+    [ForeignKey(nameof(UserId))]
+    public User User { get; set; }
 }
