@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace Item_Trading_App_REST_API.Entities
+namespace Item_Trading_App_REST_API.Entities;
+
+public class Trade
 {
-    public class Trade
-    {
-        public string TradeId { get; set; }
+    public string TradeId { get; set; }
 
-        public DateTime SentDate { get; set; }
+    public DateTime SentDate { get; set; }
 
-        public DateTime? ResponseDate { get; set; }
+    public DateTime? ResponseDate { get; set; }
 
-        public bool? Response { get; set; }
-    }
+    public bool? Response { get; set; }
 }
