@@ -4,45 +4,46 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Item_Trading_App_REST_API.HostedServices
+namespace Item_Trading_App_REST_API.HostedServices;
+
+public abstract class BaseHostedService : IHostedService
 {
-    public abstract class BaseHostedService : IHostedService
+    private readonly ILogger<BaseHostedService> logger;
+    private Timer timer;
+    private readonly TimeSpan step;
+
+    public BaseHostedService(ILogger<BaseHostedService> logger, TimeSpan step)
     {
-        private readonly ILogger<BaseHostedService> logger;
-        private Timer timer;
-        private TimeSpan step;
+        this.logger = logger;
+        this.step = step;
+    }
 
-        public BaseHostedService(ILogger<BaseHostedService> logger, TimeSpan step)
-        {
-            this.logger = logger;
-            this.step = step;
-        }
+    public Task StartAsync(CancellationToken cancellationToken)
+    {
+        timer = new Timer(async x => { await ExecuteAsync(); },
+            null,
+            TimeSpan.Zero,
+            step);
 
-        public async Task StartAsync(CancellationToken cancellationToken)
-        {
-            timer = new Timer(async x => { await ExecuteAsync(); },
-                null,
-                TimeSpan.Zero,
-                step);
-        }
+        return Task.CompletedTask;
+    }
 
-        public Task StopAsync(CancellationToken cancellationToken)
-        {
-            Log("Stopped");
+    public Task StopAsync(CancellationToken cancellationToken)
+    {
+        Log("Stopped");
 
-            return Task.CompletedTask;
-        }
+        return Task.CompletedTask;
+    }
 
-        public void Dispose()
-        {
-            timer?.Dispose();
-        }
+    public void Dispose()
+    {
+        timer?.Dispose();
+    }
 
-        protected abstract Task ExecuteAsync();
+    protected abstract Task ExecuteAsync();
 
-        protected void Log(string message)
-        {
-            logger.LogInformation("Hosted service: " + message);
-        }
+    protected void Log(string message)
+    {
+        logger.LogInformation("Hosted service: " + message);
     }
 }
