@@ -1,14 +1,9 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace Item_Trading_App_REST_API.Installers
+namespace Item_Trading_App_REST_API.Installers;
+
+public interface IInstaller
 {
-    public interface IInstaller
-    {
-        void InstallServices(IServiceCollection services, IConfiguration configuration);
-    }
+    void InstallServices(IServiceCollection services, IConfiguration configuration);
 }
