@@ -2,14 +2,13 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Item_Trading_App_REST_API.Installers
+namespace Item_Trading_App_REST_API.Installers;
+
+public class IdentityInstaller : IInstaller
 {
-    public class IdentityInstaller : IInstaller
+    public void InstallServices(IServiceCollection services, IConfiguration configuration)
     {
-        public void InstallServices(IServiceCollection services, IConfiguration configuration)
-        {
-            services.AddScoped<IIdentityService, IdentityService>();
-            services.AddScoped<IRefreshTokenService, RefreshTokenService>();
-        }
+        services.AddScoped<IIdentityService, IdentityService>();
+        services.AddScoped<IRefreshTokenService, RefreshTokenService>();
     }
 }
