@@ -1,13 +1,12 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Item_Trading_App_REST_API.Installers
+namespace Item_Trading_App_REST_API.Installers;
+
+public class MvcInstaller : IInstaller
 {
-    public class MvcInstaller : IInstaller
+    public void InstallServices(IServiceCollection services, IConfiguration configuration)
     {
-        public void InstallServices(IServiceCollection services, IConfiguration configuration)
-        {
-            services.AddControllers();
-        }
+        services.AddControllers();
     }
 }
