@@ -1,13 +1,12 @@
 ﻿using Item_Trading_App_REST_API.Models.Base;
 
-namespace Item_Trading_App_REST_API.Models.Inventory
-{
-    public class LockItemResult : BaseResult
-    {
-        public string UserId { get; set; }
+namespace Item_Trading_App_REST_API.Models.Inventory;
 
-        public string ItemId { get; set; }
-    
-        public int Quantity { get; set; }
-    }
+public record LockItemResult : BaseResult
+{
+    public string UserId { get; set; }
+
+    public string ItemId { get; set; }
+
+    public int Quantity { get; set; }
 }
