@@ -1,11 +1,12 @@
-﻿namespace Item_Trading_App_REST_API.Models.Item
+﻿namespace Item_Trading_App_REST_API.Models.Item;
+
+public record ItemPrice
 {
-    public class ItemPrice
-    {
-        public string ItemId { get; set; }
+    public string ItemId { get; set; }
 
-        public int Quantity { get; set; }
+    public string Name { get; set; }
 
-        public int Price { get; set; }
-    }
+    public int Quantity { get; set; }
+
+    public int Price { get; set; }
 }
