@@ -1,11 +1,10 @@
 ﻿using Item_Trading_App_REST_API.Models.Base;
 
-namespace Item_Trading_App_REST_API.Models.Item
-{
-    public class ShortItemResult : BaseResult
-    {
-        public string ItemId { get; set; }
+namespace Item_Trading_App_REST_API.Models.Item;
 
-        public string ItemName { get; set; }
-    }
+public record ShortItemResult : BaseResult
+{
+    public string ItemId { get; set; }
+
+    public string ItemName { get; set; }
 }
