@@ -2,12 +2,11 @@
 using Item_Trading_App_REST_API.Models.Item;
 using System.Collections.Generic;
 
-namespace Item_Trading_App_REST_API.Models.Trade
-{
-    public class BaseTradeOffer : BaseResult
-    {
-        public string TradeOfferId { get; set; }
+namespace Item_Trading_App_REST_API.Models.Trade;
 
-        public IEnumerable<ItemPrice> Items { get; set; }
-    }
+public record BaseTradeOffer : BaseResult
+{
+    public string TradeOfferId { get; set; }
+
+    public IEnumerable<ItemPrice> Items { get; set; }
 }
