@@ -1,14 +1,13 @@
 ﻿using Item_Trading_App_REST_API.Models.Item;
 using System.Collections.Generic;
 
-namespace Item_Trading_App_REST_API.Models.Trade
+namespace Item_Trading_App_REST_API.Models.Trade;
+
+public record CreateTradeOffer
 {
-    public class CreateTradeOffer
-    {
-        public string SenderUserId { get; set; }
+    public string SenderUserId { get; set; }
 
-        public string TargetUserId { get; set; }
+    public string TargetUserId { get; set; }
 
-        public IEnumerable<ItemPrice> Items { get; set; }
-    }
+    public IEnumerable<ItemPrice> Items { get; set; }
 }
