@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace Item_Trading_App_REST_API.Models.Trade
+namespace Item_Trading_App_REST_API.Models.Trade;
+
+public record RejectTradeOfferResult : TradeOfferResult
 {
-    public class RejectTradeOfferResult : TradeOfferResult
-    {
-        public string SenderId { get; set; }
+    public string SenderId { get; set; }
 
-        public string SenderName { get; set; }
+    public string SenderName { get; set; }
 
-        public DateTime ReceivedDate { get; set; } = DateTime.Now;
+    public DateTime ReceivedDate { get; set; } = DateTime.Now;
 
-        public DateTime ResponseDate { get; set; } = DateTime.Now;
-    }
+    public DateTime ResponseDate { get; set; } = DateTime.Now;
 }
