@@ -1,9 +1,8 @@
-﻿namespace Item_Trading_App_REST_API.Models.Trade
-{
-    public class RequestTradeOffer
-    {
-        public string TradeOfferId { get; set; }
+﻿namespace Item_Trading_App_REST_API.Models.Trade;
 
-        public string UserId { get; set; }
-    }
+public record RequestTradeOffer
+{
+    public string TradeOfferId { get; set; }
+
+    public string UserId { get; set; }
 }
