@@ -1,9 +1,8 @@
-﻿namespace Item_Trading_App_REST_API.Models.Trade
-{
-    public class TradeItem
-    {
-        public string ItemId { get; set; }
+﻿namespace Item_Trading_App_REST_API.Models.Trade;
 
-        public int Quantity { get; set; }
-    }
+public record TradeItem
+{
+    public string ItemId { get; set; }
+
+    public int Quantity { get; set; }
 }
