@@ -1,17 +1,16 @@
 ﻿using System;
 
-namespace Item_Trading_App_REST_API.Models.Trade
+namespace Item_Trading_App_REST_API.Models.Trade;
+
+public record TradeOffer : BaseTradeOffer
 {
-    public class TradeOffer : BaseTradeOffer
-    {
-        public string SenderId { get; set; }
+    public string SenderId { get; set; }
 
-        public string ReceiverId { get; set; }
+    public string ReceiverId { get; set; }
 
-        public DateTime SentDate { get; set; }
+    public DateTime SentDate { get; set; }
 
-        public DateTime? ResponseDate { get; set; }
+    public DateTime? ResponseDate { get; set; }
 
-        public bool? Response { get; set; }
-    }
+    public bool? Response { get; set; }
 }
