@@ -1,9 +1,8 @@
 ﻿using Item_Trading_App_REST_API.Models.Base;
 
-namespace Item_Trading_App_REST_API.Models.Trade
+namespace Item_Trading_App_REST_API.Models.Trade;
+
+public record TradeOfferResult : BaseResult
 {
-    public class TradeOfferResult : BaseResult
-    {
-        public string TradeOfferId { get; set; }
-    }
+    public string TradeOfferId { get; set; }
 }
