@@ -1,11 +1,10 @@
-﻿namespace Item_Trading_App_REST_API.Options
+﻿namespace Item_Trading_App_REST_API.Options;
+
+public record SwaggerOptions
 {
-    public class SwaggerOptions
-    {
-        public string JsonRoute { get; set; }
+    public string JsonRoute { get; set; }
 
-        public string Description { get; set; }
+    public string Description { get; set; }
 
-        public string UIEndPoint { get; set; }
-    }
+    public string UIEndPoint { get; set; }
 }
