@@ -1,70 +1,58 @@
 ﻿using Item_Trading_App_REST_API.Models.Inventory;
 using Item_Trading_App_REST_API.Models.Item;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Item_Trading_App_REST_API.Services.Inventory
+namespace Item_Trading_App_REST_API.Services.Inventory;
+
+public interface IInventoryService
 {
-    public interface IInventoryService
-    {
-        /// <summary>
-        /// Tells if the user has the quantity of the item
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="itemId"></param>
-        /// <param name="quantity"></param>
-        /// <returns></returns>
-        bool HasItem(string userId, string itemId, int quantity);
+    /// <summary>
+    /// Tells if the user has the quantity of the item
+    /// </summary>
+    Task<bool> HasItemAsync(HasItem model);
 
-        /// <summary>
-        /// Adds the quantity of the item to the user
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="itemId"></param>
-        /// <param name="quantity"></param>
-        /// <returns></returns>
-        Task<QuantifiedItemResult> AddItemAsync(string userId, string itemId, int quantity);
+    /// <summary>
+    /// Adds the quantity of the item to the user
+    /// </summary>
+    Task<QuantifiedItemResult> AddItemAsync(AddItem model, bool notify = false);
 
-        /// <summary>
-        /// Drops the quantity of the item from the inventory of the user
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="itemId"></param>
-        /// <param name="quantity"></param>
-        /// <returns></returns>
-        Task<QuantifiedItemResult> DropItemAsync(string userId, string itemId, int quantity);
+    /// <summary>
+    /// Drops the quantity of the item from the inventory of the user
+    /// </summary>
+    Task<QuantifiedItemResult> DropItemAsync(DropItem model, bool notify = false);
 
-        /// <summary>
-        /// Returns the quantity of the item of the given user
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="itemId"></param>
-        /// <returns></returns>
-        Task<QuantifiedItemResult> GetItemAsync(string userId, string itemId);
+    /// <summary>
+    /// Returns the quantity of the item of the given user
+    /// </summary>
+    Task<QuantifiedItemResult> GetItemAsync(GetUsersItem model);
 
-        /// <summary>
-        /// Returns a list of all items from the user's inventory
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <returns></returns>
-        Task<ItemsResult> ListItemsAsync(string userId);
+    /// <summary>
+    /// Returns a list of all items from the user's inventory
+    /// </summary>
+    Task<ItemsResult> ListItemsAsync(ListItems model);
 
-        /// <summary>
-        /// Locks a given quantity of the item from the user's inventory
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="itemId"></param>
-        /// <param name="quantity"></param>
-        /// <returns></returns>
-        Task<LockItemResult> LockItemAsync(string userId, string itemId, int quantity);
+    /// <summary>
+    /// Locks a given quantity of the item from the user's inventory
+    /// </summary>
+    Task<LockItemResult> LockItemAsync(LockInventoryItem model, bool notify = false);
 
-        /// <summary>
-        /// Unlocks a given quantity of the item from the user's inventory
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="itemId"></param>
-        /// <param name="quantity"></param>
-        /// <returns></returns>
-        Task<LockItemResult> UnlockItemAsync(string userId, string itemId, int quantity);
-    }
+    /// <summary>
+    /// Unlocks a given quantity of the item from the user's inventory
+    /// </summary>
+    Task<LockItemResult> UnlockItemAsync(LockInventoryItem model, bool notify = false);
+
+    /// <summary>
+    /// Returns the amount of the item with the given itemId for the user with the given id
+    /// </summary>
+    Task<LockedItemAmountResult> GetLockedAmount(GetUsersItem model);
+
+    /// <summary>
+    /// Returns the user ids that own the item with the given itemId
+    /// </summary>
+    Task<OwnedItemByUsers> GetUsersThatOwnThisItem(string itemId);
+
+    /// <summary>
+    /// Sends the notifications and clears the cached used by the deleted item
+    /// </summary>
+    Task RemoveItemCacheAsync(RemoveItemFromUsers model);
 }
