@@ -1,381 +1,543 @@
-﻿using Item_Trading_App_REST_API.Data;
+﻿using Item_Trading_App_REST_API.Constants;
+using Item_Trading_App_REST_API.Data;
 using Item_Trading_App_REST_API.Entities;
+using Item_Trading_App_REST_API.Extensions;
 using Item_Trading_App_REST_API.Models.Inventory;
 using Item_Trading_App_REST_API.Models.Item;
-using Item_Trading_App_REST_API.Services.Item;
+using Item_Trading_App_REST_API.Requests.Item;
+using Item_Trading_App_REST_API.Services.Cache;
+using Item_Trading_App_REST_API.Services.Notification;
+using MediatR;
+using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Item_Trading_App_REST_API.Services.Inventory
+namespace Item_Trading_App_REST_API.Services.Inventory;
+
+public class InventoryService : IInventoryService
 {
-    public class InventoryService : IInventoryService
+    private readonly DatabaseContext _context;
+    private readonly INotificationService _notificationService;
+    private readonly ICacheService _cacheService;
+    private readonly IMediator _mediator;
+
+    public InventoryService(DatabaseContext context, INotificationService notificationService, ICacheService cacheService, IMediator mediator)
     {
-        private readonly DatabaseContext _context;
-        private readonly IItemService _itemService;
+        _context = context;
+        _notificationService = notificationService;
+        _cacheService = cacheService;
+        _mediator = mediator;
+    }
 
-        public InventoryService(DatabaseContext context, IItemService itemService)
-        {
-            _context = context;
-            _itemService = itemService;
-        }
+    public async Task<bool> HasItemAsync(HasItem model)
+    {
+        if (string.IsNullOrEmpty(model.UserId) || string.IsNullOrEmpty(model.ItemId) || model.Quantity < 1)
+            return false;
 
-        public bool HasItem(string userId, string itemId, int quantity)
-        {
-            if (string.IsNullOrEmpty(userId) || string.IsNullOrEmpty(itemId) || quantity < 1)
-            {
-                return false;
-            }
+        var amount = await GetAmountOfFreeItemAsync(model.UserId, model.ItemId);
 
-            var amount = GetAmountOfFreeItem(userId, itemId);
+        return amount >= model.Quantity;
+    }
 
-            return amount >= quantity;
-        }
-
-        public async Task<QuantifiedItemResult> AddItemAsync(string userId, string itemId, int quantity)
-        {
-            if (string.IsNullOrEmpty(userId) || string.IsNullOrEmpty(itemId))
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
-
-            var item = GetItem(userId, itemId);
-
-            var itemData = await _itemService.GetItemAsync(itemId);
-
-            if (itemData == null)
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "Item not found" }
-                };
-            }
-
-            if (quantity < 0)
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "You cannot add a negative amount of an item" }
-                };
-            }
-            else if (quantity == 0)
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "You cannot add an amount of 0 in your inventory" }
-                };
-            }
-
-            if (item == null)
-            {
-                // then it means that we do not own items of this type
-
-                item = new OwnedItem { ItemId = itemId, UserId = userId, Quantity = quantity };
-
-                _context.OwnedItems.Add(item);
-            }
-            else
-            {
-                item.Quantity += quantity;
-            }
-
-            var updated = await _context.SaveChangesAsync();
-
-            if (updated == 0)
-            {
-                return new QuantifiedItemResult
-                {
-                    ItemId = item.ItemId,
-                    ItemName = itemData.ItemName,
-                    ItemDescription = itemData.ItemDescription,
-                    Quantity = GetAmountOfFreeItem(userId, item.ItemId),
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
-
+    public async Task<QuantifiedItemResult> AddItemAsync(AddItem model, bool notify = false)
+    {
+        if (string.IsNullOrEmpty(model.UserId) || string.IsNullOrEmpty(model.ItemId))
             return new QuantifiedItemResult
             {
-                ItemId = itemId,
-                ItemName = itemData.ItemName,
-                ItemDescription = itemData.ItemDescription,
-                Quantity = GetAmountOfFreeItem(userId, item.ItemId),
-                Success = true
+                Errors = new[] { "Something went wrong" }
             };
+
+        var item = await _cacheService.GetEntityAsync(
+            GetPrefix(model.UserId) + model.ItemId,
+            async (args) =>
+            {
+                var entity = await GetInventoryItemEntityAsync(model.UserId, model.ItemId);
+
+                if (entity is not null)
+                    return new InventoryItem
+                    {
+                        Id = entity.ItemId,
+                        Quantity = entity.Quantity
+                    };
+
+                return null;
+            });
+
+        var itemData = await _mediator.Send(new GetItemQuery { ItemId = model.ItemId });
+
+        if (itemData is null)
+            return new QuantifiedItemResult
+            {
+                Errors = new[] { "Item not found" }
+            };
+
+        if (model.Quantity < 0)
+            return new QuantifiedItemResult
+            {
+                Errors = new[] { "You cannot add a negative amount of an item" }
+            };
+        else if (model.Quantity == 0)
+            return new QuantifiedItemResult
+            {
+                Errors = new[] { "You cannot add an amount of 0 in your inventory" }
+            };
+
+        bool modified;
+
+        if (item is null)
+        {
+            // then it means that we do not own items of this type
+
+            modified = await _context.AddEntityAsync(new OwnedItem
+            {
+                ItemId = model.ItemId,
+                UserId = model.UserId,
+                Quantity = model.Quantity
+            });
+        }
+        else
+        {
+            item.Quantity += model.Quantity;
+
+            modified = await _context.UpdateEntityAsync(new OwnedItem
+            {
+                UserId = model.UserId,
+                ItemId = item.Id,
+                Quantity = item.Quantity
+            });
+
+            model.Quantity = item.Quantity;
         }
 
-        public async Task<QuantifiedItemResult> DropItemAsync(string userId, string itemId, int quantity)
+        await _cacheService.SetCacheValueAsync(
+            GetAmountKey(model.UserId, model.ItemId),
+            new InventoryItem { Id = model.ItemId, Quantity = model.Quantity });
+
+        if (!modified)
+            return new QuantifiedItemResult
+            {
+                Errors = new[] { "Something went wrong" }
+            };
+
+        if (notify)
+            await _notificationService.SendUpdatedNotificationToUserAsync(
+                model.UserId,
+                NotificationCategoryTypes.Inventory,
+                model.ItemId,
+                new InventoryItemQuantityNotification { AddAmount = true, Amount = model.Quantity });
+
+        return new QuantifiedItemResult
         {
-            if (string.IsNullOrEmpty(userId) || string.IsNullOrEmpty(itemId))
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
+            ItemId = model.ItemId,
+            ItemName = itemData.ItemName,
+            ItemDescription = itemData.ItemDescription,
+            Quantity = model.Quantity,
+            Success = true
+        };
+    }
 
-            var item = GetItem(userId, itemId);
-
-            if (item == null)
+    public async Task<QuantifiedItemResult> DropItemAsync(DropItem model, bool notify = false)
+    {
+        if (string.IsNullOrEmpty(model.UserId) || string.IsNullOrEmpty(model.ItemId))
+            return new QuantifiedItemResult
             {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "Item not found" }
-                };
-            }
+                Errors = new[] { "Something went wrong" }
+            };
 
-            if (quantity < 0)
+        var item = await _cacheService.GetEntityAsync(
+            GetPrefix(model.UserId) + model.ItemId,
+            async (args) =>
             {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "You cannot drop a negative amount of an item" }
-                };
-            }
-            else if (quantity == 0)
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "You cannot drop an amount of 0 from your inventory" }
-                };
-            }
+                var entity = await GetInventoryItemEntityAsync(model.UserId, model.ItemId);
 
-            var lockedItem = GetLockedItem(userId, itemId);
-
-            if (lockedItem != null)
-            {
-                int freeItems = item.Quantity - lockedItem.Quantity;
-                if (freeItems < quantity)
+                if (entity is not null)
                 {
-                    return new QuantifiedItemResult
+                    return new InventoryItem
                     {
-                        Errors = new[] { "You cannot drop more than you have" }
+                        Id = entity.ItemId,
+                        Quantity = entity.Quantity
                     };
                 }
-            }
 
-            item.Quantity -= quantity;
-            await _context.SaveChangesAsync();
+                return null;
+            });
 
+        if (model.Quantity < 0)
             return new QuantifiedItemResult
             {
-                ItemId = itemId,
-                ItemName = await _itemService.GetItemNameAsync(itemId),
-                Quantity = GetAmountOfFreeItem(userId, item.ItemId),
-                Success = true
+                Errors = new[] { "You cannot drop a negative amount of an item" }
             };
-        }
-
-        public async Task<QuantifiedItemResult> GetItemAsync(string userId, string itemId)
-        {
-            if (string.IsNullOrEmpty(userId) || string.IsNullOrEmpty(itemId))
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
-
-            int amount = GetAmountOfFreeItem(userId, itemId);
-
-            if (amount == 0)
-            {
-                return new QuantifiedItemResult
-                {
-                    Errors = new[] { "You do not own this item" }
-                };
-            }
-
+        else if (model.Quantity == 0)
             return new QuantifiedItemResult
             {
-                ItemId = itemId,
-                ItemName = await _itemService.GetItemNameAsync(itemId),
-                ItemDescription = await _itemService.GetItemDescriptionAsync(itemId),
-                Quantity = amount,
-                Success = true
+                Errors = new[] { "You cannot drop an amount of 0 from your inventory" }
             };
+
+        int freeItems = item.Quantity;
+        int lockedAmount = await GetAmountOfLockedItem(model.UserId, model.ItemId);
+
+        freeItems -= lockedAmount;
+        
+        if (freeItems < model.Quantity)
+            return new QuantifiedItemResult
+            {
+                Errors = new[] { "You cannot drop more than you have" }
+            };
+
+        item.Quantity -= model.Quantity;
+
+        bool modified;
+
+        if (item.Quantity == 0)
+        {
+            modified = await _context.RemoveEntityAsync(new OwnedItem
+            {
+                ItemId = item.Id,
+                UserId = model.UserId
+            });
+            await _cacheService.ClearCacheKeyAsync(GetAmountKey(model.UserId, model.ItemId));
+            await _cacheService.ClearCacheKeyAsync(GetLockedAmountKey(model.UserId, model.ItemId));
+        }
+        else
+        {
+            modified = await _context.UpdateEntityAsync(new OwnedItem
+            {
+                ItemId = item.Id,
+                UserId = model.UserId,
+                Quantity = item.Quantity,
+            });
+            await _cacheService.SetCacheValueAsync(GetAmountKey(model.UserId, model.ItemId), new InventoryItem { Id = model.ItemId, Quantity = item.Quantity });
+            await _cacheService.SetCacheValueAsync(GetLockedAmountKey(model.UserId, model.ItemId), lockedAmount);
         }
 
-        public async Task<ItemsResult> ListItemsAsync(string userId)
+        if (!modified)
+            return new QuantifiedItemResult
+            {
+                Errors = new[] { "Something went wrong" }
+            };
+
+        if (notify)
+            await _notificationService.SendUpdatedNotificationToUserAsync(
+                model.UserId,
+                NotificationCategoryTypes.Inventory,
+                model.ItemId,
+                new InventoryItemQuantityNotification
+                {
+                    AddAmount = false,
+                    Amount = item.Quantity
+                });
+
+        return new QuantifiedItemResult
         {
-            if (string.IsNullOrEmpty(userId))
+            ItemId = model.ItemId,
+            ItemName = await _mediator.Send(new GetItemNameQuery { ItemId = model.ItemId }),
+            Quantity = item.Quantity,
+            Success = true
+        };
+    }
+
+    public async Task<QuantifiedItemResult> GetItemAsync(GetUsersItem model)
+    {
+        if (string.IsNullOrEmpty(model.UserId) || string.IsNullOrEmpty(model.ItemId))
+            return new QuantifiedItemResult
             {
-                return new ItemsResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
+                Errors = new[] { "Something went wrong" }
+            };
 
-            var ownedItems = _context.OwnedItems.Where(oi => Equals(oi.UserId, userId))?.ToList();
+        int amount = await GetAmountOfFreeItemAsync(model.UserId, model.ItemId);
+        var lockedAmount = await GetLockedAmount(model);
 
-            if (ownedItems == null)
+        if (amount == 0 && lockedAmount.Amount == 0)
+            return new QuantifiedItemResult
             {
-                return new ItemsResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
+                Errors = new[] { "You do not own this item" }
+            };
 
+        var itemData = await _mediator.Send(new GetItemQuery { ItemId = model.ItemId });
+
+        return new QuantifiedItemResult
+        {
+            ItemId = model.ItemId,
+            ItemName = itemData.ItemName,
+            ItemDescription = itemData.ItemDescription,
+            Quantity = amount,
+            Success = true
+        };
+    }
+
+    public async Task<ItemsResult> ListItemsAsync(ListItems model)
+    {
+        if (string.IsNullOrEmpty(model.UserId))
             return new ItemsResult
             {
-                Success = true,
-                ItemsId = ownedItems.Select(oi => oi.ItemId)
+                Errors = new[] { "Something went wrong" }
+            };
+
+        var inventoryItems = await _cacheService.GetEntitiesAsync(
+            GetPrefix(model.UserId),
+            async (args) => await _context
+                .OwnedItems
+                .AsNoTracking()
+                .Where(oi => Equals(oi.UserId, model.UserId))
+                .Select(x => new InventoryItem { Id = x.ItemId, Quantity = x.Quantity })
+                .ToListAsync(),
+            true,
+            (InventoryItem item) => item.Id);
+
+        List<string> itemIds = new();
+
+        if (!string.IsNullOrEmpty(model.SearchString))
+        {
+            foreach (var inventoryItem in inventoryItems)
+            {
+                string itemName = await _mediator.Send(new GetItemNameQuery { ItemId = inventoryItem.Id });
+                
+                if (!itemName.StartsWith(model.SearchString, StringComparison.OrdinalIgnoreCase)) continue;
+
+                itemIds.Add(inventoryItem.Id);
             };
         }
-
-        public async Task<LockItemResult> LockItemAsync(string userId, string itemId, int quantity)
+        else
         {
-            if (string.IsNullOrEmpty(userId) || string.IsNullOrEmpty(itemId) || quantity < 1)
-            {
-                return new LockItemResult
-                {
-                    Errors = new[] { "Invalid input data" }
-                };
-            }
-
-            int amount = GetAmountOfFreeItem(userId, itemId);
-
-            if (amount < quantity)
-            {
-                return new LockItemResult
-                {
-                    Errors = new[] { "You do not own enough of this item" }
-                };
-            }
-
-            var lockedItem = await LockItem(userId, itemId, quantity);
-
-            if (lockedItem == null)
-            {
-                return new LockItemResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
-
-            return new LockItemResult
-            {
-                UserId = userId,
-                ItemId = itemId,
-                Quantity = quantity,
-                Success = true
-            };
+            itemIds = inventoryItems.Select(x => x.Id).ToList();
         }
 
-        public async Task<LockItemResult> UnlockItemAsync(string userId, string itemId, int quantity)
+        return new ItemsResult
         {
-            if(string.IsNullOrEmpty(userId) || string.IsNullOrEmpty(itemId) || quantity < 1)
-            {
-                return new LockItemResult
-                {
-                    Errors = new[] { "Invalid input data" }
-                };
-            }
-
-            int amount = GetAmountOfLockedItem(userId, itemId);
-
-            if(quantity > amount)
-            {
-                return new LockItemResult
-                {
-                    Errors = new[] { "Cannot unlock an amount more than you have locked" }
-                };
-            }
-
-            var lockedItem = GetLockedItem(userId, itemId);
-
-            if(lockedItem == null || lockedItem == default)
-            {
-                return new LockItemResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
-
-            amount -= quantity;
-
-            if (amount == 0)
-            {
-                _context.LockedItems.Remove(lockedItem);
-            }
-            else
-            {
-                lockedItem.Quantity = amount;
-            }
-
-            int removed = await _context.SaveChangesAsync();
-
-            if(removed == 0)
-            {
-                return new LockItemResult
-                {
-                    Errors = new[] { "Something went wrong" }
-                };
-            }
-
-            return new LockItemResult
-            {
-                ItemId = itemId,
-                UserId = userId,
-                Quantity = amount,
-                Success = true
-            };
-        }
-
-        private OwnedItem GetItem(string userId, string itemId) => _context.OwnedItems.FirstOrDefault(oi => Equals(oi.UserId, userId) && Equals(oi.ItemId, itemId));
-
-        private LockedItem GetLockedItem(string userId, string itemId) => _context.LockedItems.FirstOrDefault(oi => Equals(oi.UserId, userId) && Equals(oi.ItemId, itemId));
-
-        private int GetAmountOfFreeItem(string userId, string itemId)
-        {
-            var item = GetItem(userId, itemId);
-
-            var lockedItem = GetLockedItem(userId, itemId);
-
-            if (item == null || item == default)
-            {
-                return 0;
-            }
-
-            if (lockedItem == null || lockedItem == default)
-            {
-                return item.Quantity;
-            }
-
-            return item.Quantity - lockedItem.Quantity;
-        }
-
-        private int GetAmountOfLockedItem(string userId, string itemId)
-        {
-            var item = GetLockedItem(userId, itemId);
-
-            if(item == null || item == default)
-            {
-                return 0;
-            }
-
-            return item.Quantity;
-        }
-
-        private async Task<LockedItem> LockItem(string userId, string itemId, int quantity)
-        {
-            var lockedItem = GetLockedItem(userId, itemId);
-
-            if (lockedItem == null)
-            {
-                lockedItem = new LockedItem { UserId = userId, ItemId = itemId, Quantity = quantity };
-
-                _context.LockedItems.Add(lockedItem);
-            }
-            else
-            {
-                lockedItem.Quantity += quantity;
-            }
-
-            await _context.SaveChangesAsync();
-
-            return lockedItem;
-        }
+            Success = true,
+            ItemsId = itemIds
+        };
     }
+
+    public async Task<LockItemResult> LockItemAsync(LockInventoryItem model, bool notify = false)
+    {
+        if (string.IsNullOrEmpty(model.UserId) || string.IsNullOrEmpty(model.ItemId) || model.Quantity < 1)
+            return new LockItemResult
+            {
+                Errors = new[] { "Invalid input data" }
+            };
+
+        int amount = await GetAmountOfFreeItemAsync(model.UserId, model.ItemId);
+
+        if (amount < model.Quantity)
+            return new LockItemResult
+            {
+                Errors = new[] { "You do not own enough of this item" }
+            };
+
+        if (!await LockItem(model.UserId, model.ItemId, model.Quantity))
+            return new LockItemResult
+            {
+                Errors = new[] { "Something went wrong" }
+            };
+
+        if (notify)
+            await _notificationService.SendUpdatedNotificationToUserAsync(
+                model.UserId,
+                NotificationCategoryTypes.Inventory,
+                model.ItemId,
+                new InventoryItemQuantityNotification
+                {
+                    AddAmount = true,
+                    Amount = amount
+                });
+
+        return new LockItemResult
+        {
+            UserId = model.UserId,
+            ItemId = model.ItemId,
+            Quantity = model.Quantity,
+            Success = true
+        };
+    }
+
+    public async Task<LockItemResult> UnlockItemAsync(LockInventoryItem model, bool notify = false)
+    {
+        if(string.IsNullOrEmpty(model.UserId) || string.IsNullOrEmpty(model.ItemId) || model.Quantity < 1)
+            return new LockItemResult
+            {
+                Errors = new[] { "Invalid input data" }
+            };
+
+        int amount = await GetAmountOfLockedItem(model.UserId, model.ItemId);
+
+        if(amount == 0 || model.Quantity > amount)
+            return new LockItemResult
+            {
+                Errors = new[] { "Cannot unlock more than you have locked" }
+            };
+
+        amount -= model.Quantity;
+        bool modified;
+        var lockedItem = new LockedItem { UserId = model.UserId, ItemId = model.ItemId, Quantity = amount };
+
+        if (amount == 0)
+        {
+            modified = await _context.RemoveEntityAsync(lockedItem);
+            await _cacheService.ClearCacheKeyAsync(GetLockedAmountKey(model.UserId, model.ItemId));
+        }
+        else
+        {
+            lockedItem.Quantity = amount;
+            modified = await _context.UpdateEntityAsync(lockedItem);
+            await _cacheService.SetCacheValueAsync(GetLockedAmountKey(model.UserId, model.ItemId), lockedItem.Quantity);
+        }
+
+        if(!modified)
+            return new LockItemResult
+            {
+                Errors = new[] { "Something went wrong" }
+            };
+
+        if (notify)
+            await _notificationService.SendUpdatedNotificationToUserAsync(
+                model.UserId,
+                NotificationCategoryTypes.Inventory,
+                model.ItemId,
+                new InventoryItemQuantityNotification
+                {
+                    AddAmount = true,
+                    Amount = await GetAmountOfFreeItemAsync(model.UserId, model.ItemId),
+                });
+
+        return new LockItemResult
+        {
+            ItemId = model.ItemId,
+            UserId = model.UserId,
+            Quantity = amount,
+            Success = true
+        };
+    }
+
+    public async Task<LockedItemAmountResult> GetLockedAmount(GetUsersItem model)
+    {
+        if (string.IsNullOrEmpty(model.UserId) || string.IsNullOrEmpty(model.ItemId))
+            return new LockedItemAmountResult
+            {
+                Errors = new[] { "Invalid input data" }
+            };
+
+        int lockedAmount = await GetAmountOfLockedItem(model.UserId, model.ItemId);
+
+        var itemName = await _mediator.Send(new GetItemNameQuery { ItemId = model.ItemId });
+
+        if (string.IsNullOrEmpty(itemName))
+            return new LockedItemAmountResult
+            {
+                Errors = new[] { "Item not found" }
+            };
+
+        return new LockedItemAmountResult
+        {
+            ItemId = model.ItemId,
+            ItemName = itemName,
+            Amount = lockedAmount,
+            Success = true
+        };
+    }
+
+    public async Task<OwnedItemByUsers> GetUsersThatOwnThisItem(string itemId) =>
+        new OwnedItemByUsers
+        {
+            UserIds = await ListUsersThatOwnItemAsync(itemId),
+            ItemId = itemId
+        };
+
+    public async Task RemoveItemCacheAsync(RemoveItemFromUsers model)
+    {
+        foreach(var userId in model.UserIds)
+        {
+            await _cacheService.ClearCacheKeyAsync(GetAmountKey(userId, model.ItemId));
+            await _cacheService.ClearCacheKeyAsync(GetLockedAmountKey(userId, model.ItemId));
+        };
+        
+        await _notificationService.SendDeletedNotificationToUsersAsync(model.UserIds, NotificationCategoryTypes.Inventory, model.ItemId);
+    }
+
+    private Task<OwnedItem> GetInventoryItemEntityAsync(string userId, string itemId) => 
+        _context.OwnedItems
+            .AsNoTracking()
+            .FirstOrDefaultAsync(oi => Equals(oi.UserId, userId) && Equals(oi.ItemId, itemId));
+
+    private Task<LockedItem> GetLockedInventoryItemEntityAsync(string userId, string itemId) => 
+        _context.LockedItems
+            .AsNoTracking()
+            .FirstOrDefaultAsync(oi => Equals(oi.UserId, userId) && Equals(oi.ItemId, itemId));
+
+    private async Task<int> GetAmountOfFreeItemAsync(string userId, string itemId)
+    {
+        var item = await _cacheService.GetEntityAsync(
+            GetPrefix(userId) + itemId,
+            async (args) =>
+            {
+                var entity = await GetInventoryItemEntityAsync(userId, itemId);
+                if (entity is not null)
+                {
+                    return new InventoryItem
+                    {
+                        Id = entity.ItemId,
+                        Quantity = entity.Quantity
+                    };
+                }
+
+                return null;
+            },
+            true);
+
+        int lockedItemQuantity = await GetAmountOfLockedItem(userId, itemId);
+
+        return item?.Quantity ?? 0 - lockedItemQuantity;
+    }
+
+    private Task<int> GetAmountOfLockedItem(string userId, string itemId)
+    {
+        return _cacheService.GetEntityAsync(
+            GetLockedAmountKey(userId, itemId),
+            async (args) =>
+            {
+                var tmp = await GetLockedInventoryItemEntityAsync(userId, itemId);
+
+                return tmp?.Quantity ?? 0;
+            },
+            true);
+    }
+
+    private async Task<bool> LockItem(string userId, string itemId, int quantity)
+    {
+        bool storedInDb = _context.LockedItems
+            .AsNoTracking()
+            .FirstOrDefault(x => x.ItemId.Equals(itemId) && x.UserId.Equals(userId)) != default;
+        bool modified;
+
+        if (!storedInDb)
+        {
+            modified = await _context.AddEntityAsync(new LockedItem { UserId = userId, ItemId = itemId, Quantity = quantity });
+        }
+        else
+        {
+            int lockedAmount = await GetAmountOfLockedItem(userId, itemId);
+
+            quantity += lockedAmount;
+
+            modified = await _context.UpdateEntityAsync(new LockedItem { UserId = userId, ItemId = itemId, Quantity = quantity });
+        }
+
+        await _cacheService.SetCacheValueAsync(GetLockedAmountKey(userId, itemId), quantity);
+
+        return modified;
+    }
+
+    private Task<List<string>> ListUsersThatOwnItemAsync(string itemId) =>
+        _context.OwnedItems
+            .AsNoTracking()
+            .Where(x => x.ItemId == itemId).Select(x => x.UserId)
+            .ToListAsync();
+
+    private static string GetPrefix(string userId) => CachePrefixKeys.Inventory + userId + ":" + CachePrefixKeys.InventoryItems;
+
+    private static string GetAmountKey(string userId, string itemId) => GetPrefix(userId) + itemId;
+
+    private static string GetLockedAmountKey(string userId, string itemId) => CachePrefixKeys.Inventory + userId + ":" + CachePrefixKeys.InventoryLockedItem + itemId;
 }
