@@ -1,11 +1,8 @@
-﻿using System;
+﻿namespace Item_Trading_App_REST_API.Models.Trade;
 
-namespace Item_Trading_App_REST_API.Models.Trade
+public record CancelTradeOfferResult : TradeOfferResult
 {
-    public class CancelTradeOfferResult : TradeOfferResult
-    {
-        public string ReceiverId { get; set; }
+    public string ReceiverId { get; set; }
 
-        public string ReceiverName { get; set; }
-    }
+    public string ReceiverName { get; set; }
 }
