@@ -1,17 +1,18 @@
-using Item_Trading_App_REST_API.Installers;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Item_Trading_App_REST_API.Options;
 using Item_Trading_App_REST_API.HostedServices.Identity.RefreshToken;
 using Microsoft.Extensions.DependencyInjection;
+using Item_Trading_App_REST_API.Hubs;
+using Item_Trading_App_REST_API.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.InstallServicesInAssembly(builder.Configuration);
 
-builder.Host.ConfigureServices(services => services.AddHostedService<RefreshTokenHostedService>()); // had to run it here in order to have it executed after the rest of app has started
+builder.Services.AddHostedService<RefreshTokenHostedService>(); // had to run it here in order to have it executed after the rest of app has started
 
 var app = builder.Build();
 
@@ -19,6 +20,7 @@ app.UseCors(options =>
     options.WithOrigins(builder.Configuration["AngularClientSettings:Client_URL"].ToString())
     .AllowAnyMethod()
     .AllowAnyHeader()
+    .AllowCredentials()
 );
 
 // Configure the HTTP request pipeline.
@@ -52,9 +54,7 @@ app.UseRouting();
 
 app.UseAuthorization();
 
-app.UseEndpoints(endpoints =>
-{
-    endpoints.MapControllers();
-});
+app.MapControllers();
+app.MapHub<NotificationHub>("/hubs/notification");
 
 app.Run();
